replace @furystack/inject with @sensenet/client-utils
