+++ apps/sensenet/package.json
@@ -57,7 +57,6 @@
     "webpack-dev-server": "^3.8.0"
   },
   "dependencies": {
-    "@furystack/inject": "^4.0.2",
     "@material-ui/core": "^4.3.3",
     "@material-ui/icons": "^4.0.1",
     "@sensenet/authentication-google": "^2.0.13",
