+++ apps/sensenet/src/components/ErrorBoundary.tsx
@@ -1,4 +1,4 @@
-import { Injector } from '@furystack/inject'
+import { Injector } from '@sensenet/client-utils'
 import Button from '@material-ui/core/Button'
 import Dialog from '@material-ui/core/Dialog'
 import DialogActions from '@material-ui/core/DialogActions'
