+++ apps/sensenet/src/services/CommandProviders/HistoryCommandProvider.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@furystack/inject'
+import { Injectable } from '@sensenet/client-utils'
 import { CommandProvider, SearchOptions } from '../CommandProviderManager'
 import { CommandPaletteItem } from '../../hooks'
