+++ apps/sensenet/src/services/EventLogger.ts
@@ -1,5 +1,4 @@
-import { Injectable } from '@furystack/inject'
-import { AbstractLogger, LeveledLogEntry, LogLevel } from '@sensenet/client-utils'
+import { AbstractLogger, Injectable, LeveledLogEntry, LogLevel } from '@sensenet/client-utils'
 import { EventService } from './EventService'
 import { PersonalSettings } from './PersonalSettings'
