+++ apps/sensenet/src/services/PersonalSettings.ts
@@ -1,5 +1,4 @@
-import { Injectable } from '@furystack/inject'
-import { deepMerge, ObservableValue, tuple, LogLevel } from '@sensenet/client-utils'
+import { deepMerge, Injectable, LogLevel, ObservableValue, tuple } from '@sensenet/client-utils'
 import { GenericContent } from '@sensenet/default-content-types'
 import { PlatformDependent } from '../context'
 import { BrowseType } from '../components/content'
