+++ apps/sensenet/src/services/request-counter-service.ts
@@ -1,6 +1,5 @@
-import { Injectable, Injector } from '@furystack/inject'
 import { IDisposable } from 'monaco-editor'
-import { ScopedLogger } from '@sensenet/client-utils'
+import { Injectable, Injector, ScopedLogger } from '@sensenet/client-utils'
 
 interface RequestCounterRecord {
   all: number
