+++ apps/sensenet/src/sn-injector.ts
@@ -1,5 +1,4 @@
-import { Injector } from '@furystack/inject'
-import { VerboseConsoleLogger } from '@sensenet/client-utils'
+import { Injector, VerboseConsoleLogger } from '@sensenet/client-utils'
 import { EventLogger } from './services/EventLogger'
 
 export const snInjector = new Injector()
