+++ packages/sn-document-viewer-react/package.json
@@ -25,7 +25,6 @@
   "license": "GPL-2.0",
   "homepage": "https://sensenet.com",
   "dependencies": {
-    "@furystack/inject": "^4.0.2",
     "@material-ui/core": "^4.3.3",
     "@material-ui/icons": "^4.0.1",
     "@sensenet/client-utils": "^1.6.6",
