+++ packages/sn-hooks-react/package.json
@@ -25,7 +25,6 @@
   "license": "GPL-2.0",
   "homepage": "https://sensenet.com",
   "dependencies": {
-    "@furystack/inject": "^4.0.2",
     "@sensenet/client-core": "^2.2.2",
     "@sensenet/client-utils": "^1.6.6",
     "@sensenet/default-content-types": "^2.0.2",
